[PB-665] remove dialog and add view to request access
