+++ src/app/core/config/app.json
@@ -219,6 +219,13 @@
       "exact": false
     },
     {
+      "id": "request-access",
+      "layout": "empty",
+      "path": "/request-access",
+      "exact": true,
+      "auth": true
+    },
+    {
       "id": "not-found",
       "layout": "empty",
       "path": "",
