+++ src/app/core/config/views.ts
@@ -31,6 +31,7 @@ import TrashView from 'app/drive/views/TrashView/TrashView';
 import ButtonAuth from 'app/auth/views/Auth/ButtonAuth';
 import SignupBlog from 'app/auth/views/Auth/SignupBlog';
 import CheckoutCancelView from 'app/payment/views/CheckoutCancelView/CheckoutCancelView';
+import RequestAccess from 'app/drive/views/RequestAccess/RequestAccess';
 
 const views: Array<{
   id: string;
@@ -68,6 +69,7 @@ const views: Array<{
   { id: AppView.SharePhotosToken, component: SharePhotosView },
   { id: AppView.RedirectToApp, component: RedirectToAppView },
   { id: AppView.VerifyEmail, component: VerifyEmailView },
+  { id: AppView.RequestAccess, component: RequestAccess },
   { id: AppView.NotFound, component: NotFoundView },
 ];
