+++ src/app/store/slices/ui/index.ts
@@ -9,7 +9,6 @@ interface UISliceState {
   isFileInfoMenuOpen: boolean;
   isNameCollisionDialogOpen: boolean;
   isShareDialogOpen: boolean;
-  isRequestAccesDialogOpen: boolean;
   isCreateFolderDialogOpen: boolean;
   isDeleteItemsDialogOpen: boolean;
   isMoveItemsDialogOpen: boolean;
@@ -41,7 +40,6 @@ const initialState: UISliceState = {
   isFileInfoMenuOpen: false,
   isNameCollisionDialogOpen: false,
   isShareDialogOpen: false,
-  isRequestAccesDialogOpen: false,
   isCreateFolderDialogOpen: false,
   isDeleteItemsDialogOpen: false,
   isMoveItemsDialogOpen: false,
@@ -87,9 +85,6 @@ export const uiSlice = createSlice({
     },
     setIsShareDialogOpen: (state: UISliceState, action: PayloadAction<boolean>) => {
       state.isShareDialogOpen = action.payload;
-    },
-    setIsRequestAccesDialogOpen: (state: UISliceState, action: PayloadAction<boolean>) => {
-      state.isRequestAccesDialogOpen = action.payload;
     },
     setIsCreateFolderDialogOpen: (state: UISliceState, action: PayloadAction<boolean>) => {
       state.isCreateFolderDialogOpen = action.payload;
@@ -179,7 +174,6 @@ export const {
   setIsMoveItemsDialogOpen,
   setIsNewsletterDialogOpen,
   setIsSurveyDialogOpen,
-  setIsRequestAccesDialogOpen,
   setIsFileLoggerOpen,
   setIsFileInfoMenuOpen,
   setIsReachedPlanLimitDialogOpen,
